# PERF: ArrowExtensionArray.factorize