--- a/pandas/core/arrays/arrow/array.py
+++ b/pandas/core/arrays/arrow/array.py
@@ -547,21 +547,29 @@ class ArrowExtensionArray(OpsMixin, ExtensionArray):
         resolved_na_sentinel = resolve_na_sentinel(na_sentinel, use_na_sentinel)
         null_encoding = "mask" if resolved_na_sentinel is not None else "encode"
         encoded = self._data.dictionary_encode(null_encoding=null_encoding)
-        indices = pa.chunked_array(
-            [c.indices for c in encoded.chunks], type=encoded.type.index_type
-        ).to_pandas()
-        if indices.dtype.kind == "f":
-            indices[np.isnan(indices)] = (
+        encoded_idx_type = encoded.type.index_type
+        indices = encoded.combine_chunks().indices.cast(encoded_idx_type)
+        if pa.types.is_floating(encoded_idx_type):
+            fill_value = (
                 resolved_na_sentinel if resolved_na_sentinel is not None else -1
             )
-        indices = indices.astype(np.int64, copy=False)
+            indices = pc.fill_null(indices, fill_value)
+            indices = indices.cast(pa.int64())
+        # indices = pa.chunked_array(
+        #     [c.indices for c in encoded.chunks], type=encoded.type.index_type
+        # ).to_pandas()
+        # if indices.dtype.kind == "f":
+        #     indices[np.isnan(indices)] = (
+        #         resolved_na_sentinel if resolved_na_sentinel is not None else -1
+        #     )
+        # indices = indices.astype(np.int64, copy=False)
 
         if encoded.num_chunks:
             uniques = type(self)(encoded.chunk(0).dictionary)
         else:
-            uniques = type(self)(pa.array([], type=encoded.type.value_type))
+            uniques = type(self)(pa.array([], type=encoded_idx_type))
 
-        return indices.values, uniques
+        return indices.to_numpy(zero_copy_only=False, writable=True), uniques
 
     def reshape(self, *args, **kwargs):
         raise NotImplementedError(